requirements.txt
category-encoders==2.3.0
certifi==2021.10.8
click==8.0.3
colorama==0.4.4
cycler==0.11.0
Flask==2.0.2
fonttools==4.28.1
gunicorn==20.1.0
htmlmin==0.1.12
ImageHash==4.2.1
itsdangerous==2.0.1
Jinja2==3.0.3
kiwisolver==1.3.2
MarkupSafe==2.0.1
matplotlib==3.5.0
missingno==0.5.0
numpy==1.21.4
pandas-profiling==3.1.0
phik==0.12.0
Pillow==8.4.0
pydantic==1.8.2
pyparsing==2.4.7
setuptools-scm==6.3.2
tomli==1.2.2
Werkzeug==2.0.2
wincertstore==0.2
